change UTF chars for Slavonic to be rendered correctly
Same problem as a Greek passage that we had to change in the previous
commit.

Also, there was a UTF character (<U+2028>) -- which showed up several
times in the Slavonic transliteration -- that looked like a space in
the editor but was rendered as an "empty box" in HTML.  Replaced those
characters with standard spaces.